# Sorting speed improvements in SegmentInMemory 