This pr introduces 4 different changes each of which speeding up a particular part of the sorting.

Each of the changes is in a separate commit and can be looked at on it's own.

The changes are:
- Improve create_jive_table by precomputing all column values we need to sort by
- Improve create_jive_table by optimizing away unneeded sorting
- Improve sort_external by replacing an unneeded bm::bvector<> which is slow to set
- Rewrite sort_external for sparse columns to improve performance drastically

The total speed improvements are as follows (individual changes can be seen in commit messages):

Before this pr:
```
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
BM_sort_shuffled/100000/100           169 ms          156 ms           10
BM_sort_shuffled/1000000/1            659 ms          608 ms           10
BM_sort_ordered/100000/100           36.6 ms         33.8 ms           10
BM_sort_sparse/100000/100             823 ms          760 ms           10
```

After this pr:
```
-----------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations
-----------------------------------------------------------------------------
BM_sort_shuffled/100000/100_mean         92.1 ms         85.0 ms            3
BM_sort_shuffled/1000000/1_mean           173 ms          160 ms            3
BM_sort_ordered/100000/100_mean          7.26 ms         6.70 ms            3
BM_sort_sparse/100000/100_mean            110 ms          102 ms            3
```

#### Reference Issues/PRs
<!--Example: Fixes #1234. See also #3456.-->

#### What does this implement or fix?

#### Any other comments?

#### Checklist

<details>
  <summary>
   Checklist for code changes...
  </summary>
 
 - [ ] Have you updated the relevant docstrings, documentation and copyright notice?
 - [ ] Is this contribution tested against [all ArcticDB's features](../docs/mkdocs/docs/technical/contributing.md)?
 - [ ] Do all exceptions introduced raise appropriate [error messages](https://docs.arcticdb.io/error_messages/)?
 - [ ] Are API changes highlighted in the PR description?
 - [ ] Is the PR labelled as enhancement or bug so it appears in autogenerated release notes?
</details>

<!--
Thanks for contributing a Pull Request to ArcticDB! Please ensure you have taken a look at:
 - ArcticDB's Code of Conduct: https://github.com/man-group/ArcticDB/blob/master/CODE_OF_CONDUCT.md
 - ArcticDB's Contribution Licensing: https://github.com/man-group/ArcticDB/blob/master/docs/mkdocs/docs/technical/contributing.md#contribution-licensing
-->